added dataset base class
